requirements.txt
Pillow
click
numpy
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
mediapipe==0.10.0

# Tested with the following torch packages, but you should probably go to
# https://pytorch.org/ and install the latest appropriate build of PyTorch for your system
#torch==2.0.0
#torchvision==0.15.1
